Fixed bugs in "main_process.cpp"
The program suffered a segmentation fault when the elements of the vector of the results of the string analysys were accessed. I used the "std::string::at()" method instead of the"[]" operator to get the values of the elements of the vector and  "std::vector::push_back()" instead of "vector::[] = [...]" to assign elements.