Attend callbacks while scanning and new pose representation
Fixed the bug that will cause the node not to update the area values during the scan - TEST PENDING

New poses representation: Instead of 7 independent vectors, they are now a 7x4 matrix with the 7 poses. The 4 elements of each pose are the X and Y position and the components Z and W of the orientation quaternion, respectively. Each component was updated to solve the bug by which the robot could not archive the goals (we suppose the old poses would make the robot crash with the obstacles) - TEST PENDING