Make the file compilable in Kinetic (OpenCV3)
Removed the opencv2/highgui dependendence and the instructions that showed the received image as well as the thresholded ones (their only purpose was to help debugging). It compiles, but still needs TEST

NOTE: The package was developed in ROS Indigo, which use OpenCV2 by default, while Kinetic uses OpenCV3. This was the cause that didn't allow to compile the package in Kinetic and the things removed in this commit were the unique incompatibility between the two versions of OCV2 and OCV3.