Use "turtlebot_gazebo" launchers instead of our own
I've changed the launchfile so that it will use the amcl_demo.launch and turtlebot_world.launch files from the package "turtlebot_gazebo" instead of the ones we made. It passes some arguments so the used launchfiles use our world and map and make the turtlebot to start in the correct position with the appropiate orientation (according to how it is spawned in Gazebo).