Added instructions to run package nodes
Now the launchfile run the "area_publisher" and "goal_publisher" nodes from the package (p2_robots_moviles). The "main_process node is left away to allow the user to run or stop it execution because of its working flow (it must be executed for each object a user wants to ask for). This will allow future improvements such as attaching a way to ask for an object (web or app interface, a button, etc.).